Add files via upload
This booth design artfully reflects DPI's core values of profound product knowledge and innovation by seamlessly merging their expertise in aerosols, paint, and coatings with a visually stunning and conceptually rich environment. Inspired by DPI's colorful spray products and the timeless appeal of Vespa, the booth elegantly captures the essence of speed as an art form. The use of wall art, mimicking the technique of spray painting, adds a unique and captivating touch, mirroring DPI's commitment to quality and attention to detail. This immersive experience not only pays tribute to DPI's heritage but also serves as a dynamic platform for sharing their extensive knowledge and demonstrating the rigor of their product testing, offering attendees a holistic understanding of DPI's capabilities and commitment to excellence in the industry.